fix: remove warning

+++ src/core_move.cpp
@@ -368,7 +368,7 @@ Cell_t cm_update_position(bool is_turn)
 	auto pos_x = robot::instance()->getPositionX() * 1000 * CELL_COUNT_MUL / CELL_SIZE;
 	auto pos_y = robot::instance()->getPositionY() * 1000 * CELL_COUNT_MUL / CELL_SIZE;
 	map_set_position(pos_x, pos_y);
-	return Cell_t{pos_x, pos_y};
+	return Cell_t{(int16_t)pos_x, (int16_t)pos_y};
 }
 
 void cm_update_map()
